InputOnTop is the winner
Removes the old modes so we can focus on the inputOnTop mode.
